Fix: hardDelete 소유자 검증 추가 및 테스트 보완

+++ src/main/java/com/team3/deokhugam/service/comment/CommentServiceImpl.java
@@ -78,7 +78,9 @@ public class CommentServiceImpl implements CommentService {
   }
 
   @Override
-  public void hardDelete(UUID commentId) {
+  public void hardDelete(UUID commentId, UUID requestUserId) {
+    Comment comment = findComment(commentId);
+    comment.validateOwner(requestUserId);
     commentRepository.deleteById(commentId);
   }
